opportunity-agents/validation-agent/requirements.txt
# Validation Agent Dependencies

# Google AI SDK
google-genai>=0.2.0

# Reddit API
praw>=7.7.0

# HTTP requests
requests>=2.31.0

# HTML parsing (for G2 reviews)
beautifulsoup4>=4.12.0

# Type hints
typing-extensions>=4.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment variables
python-dotenv>=1.0.0
